pyproject.toml
[project]
name = "adiaryofsufferingneeds"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "firecrawl-py>=4.16.0",
    "google-genai==0.3.0",
    "pydantic==2.10.4",
    "tenacity",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]
